+++ extensions/docker/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Docker Changelog
 
+## [Improcement] - 2022-08-24
+
+- Added option to configure the Docker socket path
+
 ## [Improvement] - 2022-06-09
 
  - Added ability to Copy Container ID